chore(@clayui/tooltip): update Tooltip component stories
